Часть логики перенесена в файлы Pokemon.swift и AppModel.swift. Хранение массива покемонов перенесено в класс AppModel
