+++ PokemonListApp/PokemonListApp/AppModel.swift
@@ -1,0 +1,13 @@
+//
+//  AppModel.swift
+//  PokemonListApp
+//
+//  Created by Михаил Беленко on 12/08/2019.
+//  Copyright © 2019 Baltic Technology Company. All rights reserved.
+//
+
+import Foundation
+
+class AppModel {
+    public static var pokemonList : [Pokemon] = []
+}
